+++ src/package_manager.rs
@@ -1,41 +1,62 @@
 use std::io;
 use std::path::Path;
 
-/// An ordered mapping of project filenames to their associated package manager.
-/// Earlier entries will take precedence if a project matches multiple package managers.
-pub(crate) const PACKAGE_MANAGER_FILE_MAPPING: [(&str, PackageManager); 1] =
-    [("requirements.txt", PackageManager::Pip)];
+pub(crate) const SUPPORTED_PACKAGE_MANAGERS: [PackageManager; 2] =
+    [PackageManager::Pip, PackageManager::Poetry];
 
-/// Python package managers supported by the buildpack.
-#[derive(Debug)]
+#[derive(Clone, Copy, Debug, PartialEq)]
 pub(crate) enum PackageManager {
     Pip,
+    Poetry,
 }
 
-/// Determine the Python package manager to use for a project, or return an error if no supported
-/// package manager files are found. If a project contains the files for multiple package managers,
-/// then the earliest entry in `PACKAGE_MANAGER_FILE_MAPPING` takes precedence.
-pub(crate) fn determine_package_manager(
-    app_dir: &Path,
-) -> Result<PackageManager, DeterminePackageManagerError> {
-    // Until `Iterator::try_find` is stabilised, this is cleaner as a for loop.
-    for (filename, package_manager) in PACKAGE_MANAGER_FILE_MAPPING {
-        if app_dir
-            .join(filename)
-            .try_exists()
-            .map_err(DeterminePackageManagerError::CheckFileExists)?
-        {
-            return Ok(package_manager);
+impl PackageManager {
+    pub(crate) fn name(self) -> &'static str {
+        match self {
+            PackageManager::Pip => "pip",
+            PackageManager::Poetry => "Poetry",
         }
     }
 
-    Err(DeterminePackageManagerError::NoneFound)
+    pub(crate) fn packages_file(self) -> &'static str {
+        match self {
+            PackageManager::Pip => "requirements.txt",
+            PackageManager::Poetry => "poetry.lock",
+        }
+    }
+}
+
+/// Determine the Python package manager to use for a project, or return an error if either
+/// multiple supported package manager files are found, or none are.
+pub(crate) fn determine_package_manager(
+    app_dir: &Path,
+) -> Result<PackageManager, DeterminePackageManagerError> {
+    let package_managers_found = SUPPORTED_PACKAGE_MANAGERS
+        .into_iter()
+        .filter_map(|package_manager| {
+            app_dir
+                .join(package_manager.packages_file())
+                .try_exists()
+                .map_err(DeterminePackageManagerError::CheckFileExists)
+                .map(|exists| exists.then_some(package_manager))
+                .transpose()
+        })
+        .collect::<Result<Vec<_>, _>>()?;
+
+    match package_managers_found[..] {
+        [package_manager] => Ok(package_manager),
+        [] => Err(DeterminePackageManagerError::NoneFound),
+        _ => Err(DeterminePackageManagerError::MultipleFound(
+            package_managers_found,
+        )),
+    }
 }
 
 /// Errors that can occur when determining which Python package manager to use for a project.
 #[derive(Debug)]
 pub(crate) enum DeterminePackageManagerError {
     CheckFileExists(io::Error),
+    MultipleFound(Vec<PackageManager>),
     NoneFound,
 }
 
@@ -45,17 +66,32 @@ mod tests {
 
     #[test]
     fn determine_package_manager_requirements_txt() {
-        assert!(matches!(
-            determine_package_manager(Path::new("tests/fixtures/pip_editable_git_compiled"))
-                .unwrap(),
+        assert_eq!(
+            determine_package_manager(Path::new("tests/fixtures/pip_basic")).unwrap(),
             PackageManager::Pip
+        );
+    }
+
+    #[test]
+    fn determine_package_manager_poetry_lock() {
+        assert_eq!(
+            determine_package_manager(Path::new("tests/fixtures/poetry_basic")).unwrap(),
+            PackageManager::Poetry
+        );
+    }
+
+    #[test]
+    fn determine_package_manager_multiple() {
+        assert!(matches!(
+            determine_package_manager(Path::new("tests/fixtures/pip_and_poetry")).unwrap_err(),
+            DeterminePackageManagerError::MultipleFound(found) if found == [PackageManager::Pip, PackageManager::Poetry]
         ));
     }
 
     #[test]
     fn determine_package_manager_none() {
         assert!(matches!(
-            determine_package_manager(Path::new("tests/fixtures/empty")).unwrap_err(),
+            determine_package_manager(Path::new("tests/fixtures/pyproject_toml_only")).unwrap_err(),
             DeterminePackageManagerError::NoneFound
         ));
     }